Refactored name of Barcode classes to Barcode128 and created package.
